requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.0.5
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djongo==1.3.4
gunicorn==20.1.0
Naked==0.1.31
numpy==1.21.1
pycparser==2.20
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.4
uWSGI==2.0.19.1
whitenoise==5.2.0
